Improving phone display of full list of talks
Added new abbreviatures Ol for on-line and ZIB for Zuse Institut Berlin. Affiliations should be kept to FU, TU and HU when dealing with Berlin universities. Also improved phone display of dates in the list, to reduce the width of the dates column in it.

Old organizers should not display affiliations.